voice_assistant/requirements.txt
pvporcupine
vosk
sounddevice
numpy
pyyaml
python-dotenv
aiohttp
groq
google-generativeai
webrtcvad
piper-phonemize
onnxruntime
